refactoring: перенес store в отдельный файл optimize: создал хуки useSelector, useDispatch
